[AST] Move GenericSignatureKey to ASTContext.cpp
